Move train filter into preprocessing.
